feat: adding a difficulty factor to increase the health of enemy by 5% every day

+++ Assets/Resources/Scripts/Enemy/EnemyStats/EnemyStats.cs
@@ -1,7 +1,7 @@
 ﻿public abstract class EnemyStats
 {
 
-    public int baseHealth { get; set; }
+    public float baseHealth { get; set; }
     public float baseSpeed { get; set; }
     public float baseAttackDamage { get; set; }
     public float baseAttackRange { get; set; }
